--- src/components/Icon/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Icon";
-export type { IconProps, IconName } from "./Icon";
